separating different pages from search
